Modelling P(w|n,theta)
This file needs to be run first to generate the (PW_given_n_and_theta.mat) file, which is required for running the main driver program (compensatory_var_rescue_performace.m).
This file has the implementation for the mathematical model of compensatory variability of KC input weights given, their number of inputs and spiking threshold. Refer to the Cyan bars model in Fig5,6&7 in the publication.